添加安全出行，拍照片

+++ LookFor/LookForCommon/LookForCustomViews/RTLabel.h
@@ -32,7 +32,7 @@ typedef enum
 	NSMutableArray *_textComponents;
 	CGSize _optimumSize;
 	CGFloat _lineSpacing;
-	int currentSelectedButtonComponentIndex;
+	NSInteger currentSelectedButtonComponentIndex;
 	NSDictionary *linkAttributes;
     NSDictionary *selectedLinkAttributes;
 	id<RTLabelDelegate> delegate;
